Extract Firebase initialization into centralized module
